Only show an icon button to copy the asset ID
